+++ Drop-In/src/test/java/com/braintreepayments/api/DropInLauncherUnitTest.java
@@ -1,0 +1,24 @@
+package com.braintreepayments.api;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.robolectric.RobolectricTestRunner;
+
+@RunWith(RobolectricTestRunner.class)
+public class DropInLauncherUnitTest {
+
+    @Test
+    public void registerForActivityResult_withFragment_callsRegisterForActivityResult() {
+
+    }
+
+    @Test
+    public void registerForActivityResult_withActivity_callsRegisterForActivityResult() {
+
+    }
+
+    @Test
+    public void launchWithDropInClient_callsLaunch() {
+
+    }
+}