+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Braintree Android Drop-In Release Notes
 
 ## unreleased
-* Add support for Activity Result APIs and launching Drop-in via Fragment (fixes #313)
-* Deprecate `DropInClient#launchDropInForResult` in favor of launching Drop-in with a `DropInActivityContract`
+* Add `DropInLauncher` to support Activity Result APIs and launching Drop-in via Fragment (fixes #313)
+* Deprecate `DropInClient#launchDropInForResult` in favor of launching Drop-in with `DropInLauncher`
 
 ## 6.0.0
 * Bump braintree_android module dependency versions to `4.7.0`